Rename integration tests to make sequential nature clear
