+++ integration-tests/10_registerPlusMany.ts
